Introduce line unfolding and refolding in <screen> output for HTML
The idea is that when you break long lines in <screen> at a space character, the reader can double-click to collapse folded lines to a single line for easy copying into a terminal window, and then double-click again to refold the lines for easy reading. Line continuations with \ would allow folded lines to be used directly, at least on *n*x, and also be syntactically correct. Yet, its harder to edit a command that used \ to continue lines.

The JavaScript expects line continuations where continued lines start with a space. Lines without initial spaces do not get collapsed. For example:

<screen>$ curl
 --user admin:admin
 -X PUT
 -d '{
 "name":"joe",
 "firstname":"joe",
 "lastname":"smith",
 "email":"joe@abc.com",
 "userPassword":"11111111"}'
 http://localhost:8080/openidm/managed/user/joe
{"_rev":"0","_id":"joe"}</screen>

Shows up as:

$ curl
 --user admin:admin
 -X PUT
 -d '{
 "name":"joe",
 "firstname":"joe",
 "lastname":"smith",
 "email":"joe@abc.com",
 "userPassword":"11111111"}'
 http://localhost:8080/openidm/managed/user/joe
{"_rev":"0","_id":"joe"}

And then when you double-click:

$ curl --user admin:admin -X PUT -d '{ "name":"joe", "firstname":"joe", "lastname":"smith", "email":"joe@abc.com", "userPassword":"11111111"}' http://localhost:8080/openidm/managed/user/joe
{"_rev":"0","_id":"joe"}

And vice versa.

There's no doubt cleanup to do in the core doc sources to take advantage of this.

git-svn-id: https://svn.forgerock.org/openidm/trunk@551 d98387aa-ee2c-4292-a9e6-504d2a719fd3