OpenIDM Fix reauth policy to not fail when value has not changed http://sources.forgerock.org/cru/CR-1057
Calls to patch were failing now that validation is properly being applied to patch calls. The failure was due to the fact that the whole object was being validated, and some properties in the object were requiring reauth. So even if you were not changing a property which required reauth, the reauth validation rule was being raised. This fixed the problem by not raising the failure if the value being evaluated is the same as the current value for the object. 

https://bugster.forgerock.org/jira/browse/OPENIDM-1011

git-svn-id: https://svn.forgerock.org/openidm/trunk@1739 d98387aa-ee2c-4292-a9e6-504d2a719fd3