OPENIDM-2106 (CR-4210) Refactor AuditService log filters from using a map of eventtype to some set of actions to a collection of log filters with a defined interface.  Calling
    filter.isFiltered(context, request)

lets AuditSerivceImpl#handleCreate determine whether to filter the request.
This allows extension to other types of filters but also lets different event types
use a different set of action constants and thus solves OPENIDM-2106.

All AuditLogFilter implementation is hidden behind the static factory methods in
AuditLogFilters.

Note that in the implementations of newActivityActionFilter and newReconActionFilter,
the resultant filter does not filter nulls or actions that are not constants of the
requisite enum type.



git-svn-id: https://svn.forgerock.org/openidm/trunk@3792 d98387aa-ee2c-4292-a9e6-504d2a719fd3